added rle example
I think that is all the example recursions I will add.  We have then
just the derivation, interface for subsidiary induction, and examples
of subsidiary induction left to write, and filling in the related
work.  Space is getting tight, so we might need to shrink some earlier
parts a bit.