decomposed swap into swap_elt and map
I was struggling with the proof of swap_break (Program_Playground) for
our version of break, and was thinking it might help manage some
complexity if I change the definition of swap.  I am not sure it is
helping.  While intellectually appealing, it does come with the price,
at least with my level of ability in Coq, of some extra steps to
unfold swap_elt in various places.  Probably I am doing something
suboptimal there...