ok, proved the equivalent of swap_break and swap_break_snd
So far, this is not showing any clear advantage over Program,
and arguably it is a bit worse: the proof is a somewhat longer, due
to the need to deal with the trick used in writing span based on spanh
(in Span.v).  And that trick arises due to the approach.  Although
actually we should revisit the code for Span using our
divide-and-conquer recursion, because it would get simplified.